+++ src/polkadot/info/liberlandInfo.ts
@@ -26,7 +26,7 @@ const networkInfo: PolkadotNetworkInfo = {
     'wss://liberland-rpc.dwellir.com/'
   ],
   ss58Format: 42,
-  subscanBaseUrl: undefined,
+  subscanBaseUrls: [],
   subscanQueryLimit: 100,
   partialFeeOffsetMultiplier: '2',
   lengthFeePerByte: '76800000',