+++ src/polkadot/info/polkadotInfo.ts
@@ -9,9 +9,12 @@ import {
 } from '../polkadotTypes'
 
 const networkInfo: PolkadotNetworkInfo = {
-  rpcNodes: ['wss://rpc.polkadot.io'],
+  rpcNodes: ['wss://polkadot-asset-hub-rpc.polkadot.io'],
   ss58Format: 0,
-  subscanBaseUrl: 'https://polkadot.api.subscan.io/api',
+  subscanBaseUrls: [
+    'https://polkadot.api.subscan.io/api',
+    'https://assethub-polkadot.api.subscan.io/api'
+  ],
   subscanQueryLimit: 100,
   partialFeeOffsetMultiplier: '1',
   lengthFeePerByte: '1000000',